+++ packages/unixfs/src/commands/ls.ts
@@ -1,9 +1,10 @@
-import { exporter, type UnixFSEntry } from 'ipfs-unixfs-exporter'
-import mergeOpts from 'merge-options'
+import { mergeOptions as mergeOpts } from '@libp2p/utils/merge-options'
+import { exporter } from 'ipfs-unixfs-exporter'
 import { NoContentError, NotADirectoryError } from '../errors.js'
 import { resolve } from './utils/resolve.js'
 import type { LsOptions } from '../index.js'
 import type { GetStore } from '../unixfs.js'
+import type { UnixFSEntry } from 'ipfs-unixfs-exporter'
 import type { CID } from 'multiformats/cid'
 
 const mergeOptions = mergeOpts.bind({ ignoreUndefined: true })